Compare AWS S3 header MD% to downloaded file MD5.
Added a check to compare the value in the md5 field of the aws header to
the computed MD5 of the downloaded file. If they do not match an error is
logged and the file is downloaded again. If the MD5 sums still do not
match another error is logged and the second download attempt is returned.