Controller for REST api requests
The controller parses the id part of the request into the bank id and item
id by splitting the request value on the '-' character to match the sample
API request. The codes are optional. If no codes are specified an empty array is
used to denote the lack of codes so the codes value is not null.