Include iris (#31)
* add iris and dependencies

Overlay Iris on the project. Downgrade some dependencies so they work with
Iris. Removed Spring Boot because it does not work with overlays. Add the
logging method used by Iris.

* Basic controller for items

The controller currently only renders a hardcoded item. THis is intended.
This will be changed to render the item in the URL.

* Models for item and accommodations

Model the item and its accommodations into an object so that it can be shaped into a JSON
object that can be passed off to the JavaScript frontend.

* Set language level

We need to use Java 7 for this project.

* Remove unneeded JSON annotations

I changed the way I was building the JSON token that is passed to the
javascript frontend so it gets its own object. The JSON annotations on the
item request were no longer needed.

* Remove Java 8 only method to convert long to int

We need to use Java 7 for this project so the very useful Math.toIntExact
function that converts from a long to an int is no longer available.

* Wait until blackbox loads to show the item.

The blackbox needs to fully load before the accommodaions can be set.
Wait to load the item until after the Blacbox has signaled that it has
finished loading so we can load the item correctly.

* Added JavaDocs

* Use itemvierwserservice web.xml

There is no easy way to merge the web.xmls for the iris verlay andthe
itemviwerservice. The itemviewerservice web.xml was overrwiting the one
from the iris overlay. It is clearer to future devs, and prevents unneeded
imports if we do not import the web.xml from the iris overlay.

* Set the iframe to scale with the page.

We are using an iFrame to display the content and we need use our own
pages instead of the ones inherited from Iris so we can load the content
as soon as the page loads.

* Fixed style errors

* Fixed typo in highlight accommodation string

The typo was preventing the highlighter from working because it was not
recognized

* Change dependencies to use Java7 and slf4j logging

* Use context path for the iris iframe location

By using the context path the iframe source now uses the correct relative
path.

* Removed unused controllers

Logging will be done via the diagnostic API. That will get its own
controller at some point. This means we do not need the logs controller.
The example controller was just there for reference and is no longer
needed.