Feature diagnostic api (#34)
* Add diagnostics Api controller and core

Set up a controller to return the diagnostics API xml as text/xml per the
specifications. Added core logic to run the diagnostics.

* Add diagnostic classes and dependencies

I created a base class for the diagnostics and extended it so we can
easily add diagnostic checks or extend existing ones. I am using the
built in Java XML binding (JAXB) functions to serialize the XML because
they make it easy to add or remove diagnostic tests.

* Log errors the diagnostic API finds

When the diganostic API finds an error also log it to the system logs for
easier debugging.

* Exclude slf4j-log4j12 jar from iris.war

There was a conflict with slf4j between log4j12 and logback-classic, both
of which were in Iris, and were preventing the logging from working
correctly. I settled on logback-classic because it is what the TDS and AIR
packages use. log4j12 is used in some of the packages they depend on and
should have been excluded at that point. I do not havecontrol over those
packages so I'm excluding it here.

* Add configuration and properly log with logback

After fixing the logging configuration conflicts I added the required
configuration files and dependencies to correctly use logback-classic in
the item viewer service. I'm using it because it seems to work well, and
it is used in Iris and the other TDS packages.

* Remove unused class for tracking diagnostic results

I'm using a base diagnostic class and extensions of that class to run
diagnostics, then serializing them with JAXB. The DiagnosticResult class
is not needed to track the diagnostic results any more so I am removing
it.

* Return 500 error on total diagnostic failure

If there is a server error when attempting to run the diagnostics return a
500 status.

* Refactor diagnostic API class names

I refactored the diagnostic API classes to reflect what they are testing
instead of just the diagnostic level. If we ever change what levels a
specific diagnostic is called at, or insert more levels this will result
in less refactoring in the future. It also makes it easier to tell what a
diagnostic test is doing.

* Update class names to reflect the refactored names

* Add default value for status level

The diagnostic status level no longer has to be specified. If it is not
specified then it defaults to level 0.

* Change how diagnostic rating is calculated

The overall diagnostic rating should fail if any of the 5 levels of tests
fail. I changed how the diagnostics are run and the rating is calculated
to separate out class instantiation, test running and result generation.

* Refactored database diagnostics into one class

The database diagnostics were split across level 3 and level 4
and different classes. I refactored them into one class that can be used
to run the database category of diagnostics.

* Add class and dependencies for testing

* Changes to make the classes more testable

As I was writing tests for the diagnostic API I made changes to the
classes to make them easier to test or fix small issues I found.

* Unit tests for diagnostic classes

Added a unit test for the diagnostic classes excepting the system
diagnostics which are not implemented and the external providers
diagnostics which does not have much internal logic.

* Fixed bug found in testing

Due to the way I was checking ifa list was null or empty it could throw a
null pointer exception because the empty check came before the null check.
Changing the order of the checks resolves the issue.

* fixed test logic issues

Some of the tests were Asserting in the incorrect order so the expected
value was second. There were a few other minor errors such as checking
against the wrong status.

* Add system level diagnostics and dependencies

I am using the OSHI library to fetch system information for the system level
diagnostics. This required me to exclude an older version of JNA from being
included with Iris because of dependency conflicts.

* Style fixes and JavaDocs

Added Javadocs comments where they were needed. Fixed code style issues such as
incorrect whitespace.