Remove redis (#35)
* Removed Redis code

Removed the Redis code because we are no longer using it. Made some initial
replacements with code to just use the filesystem.

* download, unzip and replace content packages

When a new content package is listed in Amazon S3, or when the app is started
it will download all of the content packages from Amazon S3 into a temporary
directory. It will then unzip the content packages. Next it will remove the
contents of the iris content directory. Finally it will copy the
contents of the temporary content directory to the iris content directory. The
purpose of this is to fully update the content directory any time the content
is changes without tracking each file in the content directory.

* Reload Iris on content pack update

When new content is downloaded Iris needs to be reloaded by requesting
the reload url. The only way to force the content reload is to make a request
to that URL so I added a controller that redirectds to the Iris content
reload URL, and then I make an HTTP request to the controller.

* Use fixed delay instead of fixed rate

Fixed rate bases its start time off of when the previous start time was. This
is a problem if the task takes longer than the interval because it will
immediatly run the task again with no delay. By using fixed delay it bases the
delay off of the end time of the previous task.

* Javadocs and checkstyle violations