Return a 404 status when an item does not exist (#36)
Added a custom exception to return a 404 status when someone tries to load an
item that does not exist. Also added Javascript to display an alert if someone
makes it as far as loading an incorrect item in JavaScript. This prevents
people from getting a blank page if they try to load an item that does not
exist.